+++ android/src/main/kotlin/com/sap/gigya_flutter_plugin/GigyaFlutterPlugin.kt
@@ -70,6 +70,7 @@ class GigyaFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
             "resolveSetAccount" -> sdk.resolveSetAccount(call.arguments, result)
             "forgotPassword" -> sdk.forgotPassword(call.arguments, result)
             "initSdk" -> sdk.initSdk(call.arguments, result)
+            "getSession" -> sdk.getSession(result)
             "setSession" -> sdk.setSession(call.arguments, result)
             "sso" -> sdk.sso(call.arguments, result)
             "webAuthnLogin" -> {