+++ lib/gigya_flutter_plugin.dart
@@ -23,6 +23,7 @@ enum Methods {
   showScreenSet,
   forgotPassword,
   initSdk,
+  getSession,
   setSession,
   sso,
 }
@@ -293,6 +294,20 @@ class GigyaSdk with DataMixin, GigyaResponseMixin {
     }).timeout(getTimeout(Methods.removeConnection), onTimeout: () {
       debugPrint('A timeout that was defined in the request is reached');
       return timeoutError();
+    });
+    return result;
+  }
+
+  /// Get current session.
+  Future<Map<String, dynamic>?> getSession() async {
+    final result = await _channel.invokeMapMethod<String, dynamic>(
+      Methods.getSession.name,
+    ).catchError((error) {
+      debugPrint('get session error $error');
+      throw error;
+    }).timeout(_getTimeout(Methods.getSession), onTimeout: () {
+      debugPrint('A timeout that was defined in the request is reached');
+      return _timeoutError();
     });
     return result;
   }